backend/requirements.txt
fastapi==0.104.1
uvicorn==0.38.0
pymongo==4.6.0
motor==3.3.2
beanie==1.24.0
redis==5.0.1
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.12.0
websockets==15.0.1
crewai
litellm
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.0
aiohttp==3.8.6
openrouteservice==2.3.3
